+++ test/graph_test.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/widgets.dart';
 import 'package:graphview/GraphView.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -41,6 +42,35 @@ void main() {
       expect(graph.nodeCount(), 9);
     });
 
+    test('Node Hash Implementation is performant', () {
+      final graph  = Graph();
+
+      var rows = 1000000;
+
+      var integerNode = Node.Id(1);
+      var stringNode = Node.Id("123");
+      var stringNode2 = Node.Id("G9Q84H1R9-1619338713.000900");
+      var widgetNode = Node.Id(Text("Lovely"));
+      var widgetNode2 = Node.Id(Text("Lovely"));
+      var doubleNode = Node.Id(5.6);
+
+      var edge = graph.addEdge(integerNode, Node.Id(4));
+
+      var node = doubleNode;
+      var stopwatch = Stopwatch()..start();
+      for (var i = 1; i <= rows; i++) {
+        var hash = node.hashCode;
+
+      }
+
+      var timeTaken = stopwatch.elapsed.inMilliseconds;
+
+      print('Timetaken $timeTaken ${node}');
+
+      expect(timeTaken < 100, true);
+    });
+
 
   });
+
 }